Revert "refactor(enumReducer): removed redundant definition"
This reverts commit 4db2ab02289114a00845bce7f3acccc38ad53916.
